remove old webcomponent, give odl name the new webcomponent
